+++ RELEASE_NOTES.md
@@ -18,6 +18,7 @@
 * `PhysicsEngine` should no longer be instantiated in the Server `main.js` and in the client entry point.  Rather, it should be instantiated in the `GameEngine`.
 * `PhysicsEngine` constructor now does initialization instead of the `init` function
 * `GameEngine` step method cannot be called without passing the `isReenact` argument.  Games which override the `step` method must pass this argument when calling the super method.
+* Objects are now instantiated with a reference to the gameEngine, and get and ID automatically
 
 ## Release 1.0.1